+++ web/src/components/dashboards/addPanel/DashboardQueryBuilder.vue
@@ -761,16 +761,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
                             style="width: 100%"
                             :rules="[(val) => !!val || 'Required']"
                           />
-                          <!-- <CommonAutoComplete
-                            v-if="
-                              !['Is Null', 'Is Not Null'].includes(
-                                fields?.operator
-                              )
-                            "
-                            :field="fields"
-                            :index="index"
-                          ></CommonAutoComplete> -->
-                          <CommonAutoComplete2
+                          <CommonAutoComplete
                             v-if="
                               !['Is Null', 'Is Not Null'].includes(
                                 dashboardPanelData.data.queries[
@@ -787,7 +778,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
                             :items="dashboardVariablesFilterItems"
                             searchRegex="(?:^|[^$])\$?(\w+)"
                             :rules="[(val: any) => val?.length > 0 || 'Required']"
-                          ></CommonAutoComplete2>
+                          ></CommonAutoComplete>
                         </div>
                       </q-tab-panel>
                       <q-tab-panel
@@ -916,7 +907,6 @@ import SortByBtnGrp from "@/components/dashboards/addPanel/SortByBtnGrp.vue";
 import HistogramIntervalDropDown from "@/components/dashboards/addPanel/HistogramIntervalDropDown.vue";
 import { useQuasar } from "quasar";
 import CommonAutoComplete from "@/components/dashboards/addPanel/CommonAutoComplete.vue";
-import CommonAutoComplete2 from "@/components/dashboards/addPanel/CommonAutoComplete2.vue";
 
 export default defineComponent({
   name: "DashboardQueryBuilder",
@@ -926,7 +916,6 @@ export default defineComponent({
     HistogramIntervalDropDown,
     DashboardSankeyChartBuilder,
     CommonAutoComplete,
-    CommonAutoComplete2,
   },
   props: ["dashboardData"],
   setup(props) {
